requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file base.txt base.in
#
et-xmlfile==1.0.1         # via openpyxl
jdcal==1.3                # via openpyxl
openpyxl==2.4.9
psycopg2==2.7.3.1
python-dateutil==2.6.1
pyyaml==3.12
six==1.11.0               # via python-dateutil
sqlalchemy==1.1.14
xlrd==1.1.0
